requirements.txt
# Core Dependencies
pinecone-client==5.0.1
streamlit==1.31.0
python-dotenv==1.0.0

# Document Processing
PyMuPDF==1.23.26
pypdf==4.0.1
python-docx==1.1.0

# Text Processing & Chunking
langchain
langchain-community
langsmith
langchain-text-splitters
tiktoken

# LLM Integration (Free Options)
groq==0.4.2
anthropic==0.18.1  # Optional: if user has API credits

# Utilities
pydantic
requests
numpy==1.26.3
pandas==2.2.0
